change TimePosted to Int64

+++ musicCatalogDotNetAPI/remailDotNetAPI/Controllers/ChatMessageController.cs
@@ -140,7 +140,7 @@ namespace musicCatalogDotNetAPI.Controllers
         public string Content { get; set; }
 
         [JsonPropertyName("timePosted")]
-        public int TimePosted { get; set; }
+        public Int64 TimePosted { get; set; }
 
     }
